fmt: Add printf-style formatting
The formatting that we had before was based on individual function
calls to append data to a buffer. This is simple, but becomes very
verbose quickly. Now we have a `fmt` function that uses a
printf-like syntax to format a number of arguments in a buffer.
An annoying part of this is that we have to deal with a C-string
(the `fmt` parameter of the `fmt` function), but at least formatting
becomes easy.