fmt, arena: Make `fmt` independent of `arena`
Out of all the fundamentals, we need some parts to be available
in assertions or crashes. These parts must themselves not rely
on assertions or crashes. Basically, we need this to implement
the `assert` and `crash` macros: base types, strings, formatting
and printing. This list of modules can't use assertions internally.

The arena implementation doesn't make sense if it can't use assertions.
At least that's how it's designed right now. It's designed for ease of
use and safety. Before formatting depended on arenas and thus we
couldn't use formatting in assertions. Now formatting is independet
of arenas.