fmt, print: Switch to using `struct str` for format string
We're in a dilemma. The compiler can check if printf-style format
strings are used correctly. But this is under the condition, that
we use a `char *` as the format string. Additionally, it expects
the format string to use '%s' for `char *`. We don't want to use
this kind of NULL-terminated string though. We want to use
`struct str` instead. But if we use `struct str`, we loose the
compiler's support. Using the conventional printf syntax means
using NULL-terminated strings for printing. In turn, that means
that all the `struct str` and `struct str_buf` strings need to
be NULL-terminated so that they can be printed.

The tradeoff is: start using NULL-terminated strings everywhere
or give up on compiler support to check if format strings are
correct. I opted for the latter option because strings are used
much more than formatting. Just be careful whenever you format
something.