bootloader: Add print functions to `load_kernel.c`
This commit copies code from `src/com.c` and `src/print.c` into
`bootloader/load_kernel.c`. This isn't pretty but fine as we don't
need high-quality drivers in the bootloader for now. I.e., as the
drivers improve, it's fine if the code in the bootloader stays simple.
The most important thing is that the bootloader can print basic
messages for debugging.