bootloader: Load ELF into static buffer
Picking a random address in memory to temporarily store the
ELF header and program header table is _very_ unsafe. Now we
reserve some space in the bootloader binary image where we can
load these parts of the ELF file. Small program header tables are
in the low hundreds of bytes so the buffer should have plenty of
space for now. If we need more space later, we can just bump the
size of the bootloader a little bit.