config.mk: Add new configuration system
The config.mk file is a simple config file format. It contains
assignments as can be seen in this diff. It's a valid Makefile
and can be loaded into Makefiles verbatim. There's a new script
`make_config.sh` that parses the config.mk file and generates
equivalent config files in different languages. E.g., `config.mk`
can be turned into a header `config.h` that contains all the same
definitions but for C. The generated configs live in the `build/`
directory and they can be included in NASM, linker scripts and C.

This is a way to keep common definitions in one place. The goal
is to never define the same constant twice.