print: Add `print` function that doesn't use an explicit buffer
I didn't want to do this initially, but it's just too annoying
to be forced to always specify a buffer. Especially for debugging.
And it let me to create tons of small, temporary buffers on stack
which isn't very nice either. Now we have one function, `print`,
that creates a small buffer on stack for formatting. If the buffer
is too small for the string, the formatting operation will fail.
Don't use `print` for complicated stuff. Just use it for small
debug prints.