base: Add `static_assert` macro
This is a trick in C to check the value of an expression at
compile time. The size of an array type declaration is computed
at compile time. It's an error if the size is negative. So we
can use the ternary operator to check if the condition that we
pass to the assertion is true. If it is, we declare an array
type of size 1. If it's false, we declare an array type of
size -1. This will lead to a compile error.