init: Fix IDT memory corruption bug
The way GCC currently compiles this code is that the IDT lives in
`.bss` right before the `init_kernel_stack` that's used before
memory allocation has been set up to allocate a proper kernel stack.
I made the mistake of setting `rsp` to the address `init_kernel_stack`
instead of the address `init_kernel_stack + sizeof(init_kernel_stack)`.
Because the stack grows downward, it would grow into the memory that
the IDT occupies. The bigger the stack got, the more garbage it wrote
into the IDT.

I didn't notice this issue before because I only used the first few
entries in the IDT while only the upper ones where corrupted. Ever
sinec using vector 128 for syscalls, I started randomly getting
general protection faults. The reason, of course, was that the
stack grow too big and started corrupting entry 128 in the IDT.

While we're at it, I also increased the size of the `init_kernel_stack`,
and the normal kernel stack for safe measure.