init: Fix off-by-one error in kernel stack
I think this is correct the way it is now because adding the size
of a buffer to the address of the first byte in the buffer gives
the address of the first byte past the end of the buffer. Thus, we
need to subtract 1 to get the address of the last byte in the buffer.