config: Update and document memory layout constants
We don't need quite as much space for the .entry section.
It's about 80k bytes in the ELF right now and I gave it
128KiB in the static config which should be fine. I don't
expect this section to grow much given it's only briefly
used to set up a temporary page table to jump into the kernel.
In fact, I'm surprised it's that big ... May be worth looking
into