paging: Add virt-phys mapping and fix impl
This commit does two things: first, it adds a fixed, linear
mapping between kernel virtual and physical addresses and
second, it uses the new mapping to fix the paging implementation
so that the paging implementation doesn't access physical
addresses (thus requiring identity mappings to be present)
as it did up until now.

There are two regions of memory used by the kernel, called
"kernel memory": the region for code and data where to kernel
ELF lives, and a large region of memory that's dynamically
managed.

The regions both have a fixed length, a virtual base address
and a physical base address. The first virtual address is
mapped to the first physical address. Same thing for the second
address and so forth -- it's a linear mapping.

With knowledge of this linear mapping, it becomes easy to
translate between virtual and physical addresses in both
directions. The virtual to physical direction was easy before
because one just has to walk the page table. But while a
single virtual address can only point to a single physical
address, multiple different virtual addresses can point to the
same physical address (viewed as a mathematical function, paging
is not injective so it can't be inverted).

Using the linear mapping, we do, however, gain the ability to
translate all physical addresses from kernel memory back to
their unique kernel virtual address. It may be that other virtual
addresses also lead to this physical address, e.g., if a process
is using that page. But the address translation using the linear
mapping will always return the virtual address from kernel memory.

This is convenient, because all memory used in the kernel (both
for internal data structures and for resources required by
processes) will be allocated from the dynamic region of kernel
memory. This means all memory is accessible though kernel virtual
addresses.

---

There were bugs before in the paging implementation where physical
addresses were used. I didn't catch those bugs before because I
had an identity mapping in place for the first few megabytes of
memory so it was possible to dereference the physical addresses.

We can't always have such an identity mapping for all physical
memory, though. Now the new ability to translate physical to
virtual addresses is used instead so that only virtual addresses
are dereferenced.