kvalloc: Add kernel allocator for virtual addresses
kvalloc can be used globally to allocate virtual addresses.
It will manage the dynamic region of kernel memory and be
default source of memory for all allocators.

As the dynamic region of kernel memory is mapped linearly to
physical addresses, allocators from kvalloc can also be used
when regions of contiguous physical pages are required. This
is an implementation detail of the current paging system,
though, and may theoretically change at any point in the
future.