ramfs: Update allocation strategy
The RAM fs is more flexible with allocations now and has a
single generic allocator passed to it. From that, a pool for
fs nodes is allocated (simply for efficiency). Strings and
file memory are allocated from the main allocator.

Because strings are often small, it is very wasteful to
allocate them from the main allocator. A dedicated string
pool would be valuable.