kernel.ld: Add custom PHDRS for extra ELF layout control
With the custom PHDRS, we can more tightly control how
the ELF is laid out and loaded. I want to use the `extra_seg`
segment for additional data that can be loaded by the kernel,
such as a file system image.

I want to place such extra data into a separate segment so it
is kept separate from the main parts of the ELF (text, (ro)data,
etc.).

Note that none of the permissions on these segments are enforced
at present. The `entry_seg` segment requires RWX permissions because
the bootloader modifies memory right next to its own code.