paging: Fix VAS bounds calculations and avoid double allocations
Bounds calculations: We don't mandate a page alignment for the
length of the VMA anymore. This didn't make much sense in the
first place, given the base address didn't need to be page
aligned.

Instead, the address of the first page past the VMA will be
calculated as

	vaddr_page_end = ALIGN_UP(vma.base + vma.len, PAGE_SIZE);

Then, beginning at `vma.base`, all pages up to and excluding the
page of `vaddr_page_end` will be mapped/unmapped/modified.

The advantage of this is that it allows finer-grained copies in
`vas_memcpy` because VMAs can be less that a page in size.

Double allocations: Before allocating a page to a virtual address,
we check if the is already a mapping for the address in the VAS.
Mapping the virtual address again is undesirable because everything
on the old physical page will be lost.