bootloader: Document findings on booting on real hardware
I spent the day trying to get the kernel to boot from
an USB stick on an old laptop of mine.

After running into some issues with the bootloader (see
the comments in bootloader.s), I realized I had to write
a USB driver of some kind to load the full kernel image or
re-write the bootloader to use BIOS to load the full kernel
image because I only have an ATA driver for loading at this
point.

Both options would take more time than I have now, so I will
delay this to a later point in time.