error: Make `struct_result(...)` end in semicolon
The struct_result macro defines functions and, as such,
had to be called at the global scope without a semicolon
at the end (well you could put a semicolon there, but the
compiler would complain that it doesn't belong there).

Clang-fmt got confused by the macro call without a semicolon
at the end and often messed up formatting. Now we have found
a workaround that allows (in fact, mandates) placing a
semicolon at the end.