We don't need a user space to serve web pages
Writing a full operating system that has something
interesting about it exceeds my current skills is
an endless and time-consuming project. I do not
have the motivation to purse this project forever
-- especially not without a clear goal in mind.

So I decided on a goal a while ago: to serve my
blog web site. This is a technical challenge I
can learn a lot from! It is also a great exercise
in first-principles engineering by learning how
to write exactly the low-level code requires to
complete this task, from the ground up.

Building without a set goal in mind is pointless!
It will lead to long, aimless roaming and writing
code that no one really needs. (Of course, there is
a place for experimentation.) With a goal in mind,
however, it's possible to ask the question of what
exactly needs to happen to reach this goal.

In this case, a user space does not need to happen
to serve a blog. The system will run in a VM where
isolation is provided by the hypervisor. We just need
a single (or multiple) ring 0 process, that acts as
a web server.

I have just caught myself writing a lot of pointless
code to start implementing a user space for no apparent
reason except that this is the stuff textbooks talk
about. Out of my own insecurity, I felt pressure to
implement this stuff, so I have a "real" operating
system. I have not implemented a user space before
so I'm afaid people might claim that I just couldn't
do it.

Well, I probably could, if I invested more time into this.
But from a technical point of view, it's the wrong
approach, and that's all that counts. Plus, it's better
for my own motivation to work straight towards a clear
goal.

So with this commit, I'm coming clean of the mistake
I made by beginning to implement user space stuff. It's
hard to identify what exactly is needed to solve a
specific problem. That's why many people just do the
same stuff everybody does so all code looks the same
and is equally shitty. We don't want that. So the
next time I err like this, I will come back here.