paging: Delete VAS/VMA stuff and clean up
The paging subsystem is one of the first things I wrote, when I
didn't have very clear goal in mind. Specifically, I modeled the
VAS/VMA functions after what I found in Unikraft thinking it is
a nice interface and could be useful in the future.

Well this sort of thinking is always misled, even if it's a nice
interface. Because without a set purpose, the implementation is
doomed to be nonsensical.

And so it was! It evolved a bit over time when I tried to change
it so that it works well for allocating memory for user processes
but it remained problematic. I had problems with the API all the
time and changed it on multiple occasions to actually work for
the stuff I needed. See, this is what you get from writing code
for a problem you don't yet have.

Since the user process stuff has been deleted, there is no need
for this code any more. The paging code is now much simpler and
suits my needs better.

Cheers!