manuals: Add reference manuals
This commit includes a copy of the Intel manuals that I use to
write this system. This is a great way to future-proof this
project by making sure that anyone looking at the code will
always have access to the same versions of the manuals that I
did at the time of writing.