print: Add debug levels
This helps combat debug message spam. Messages that will only be
printed on few occasions can be marked `PINFO`. Other messages
that would be printed very often and are only useful when debugging
a subsystem should be `PDBG`. Messages that should always appear
can be marked `PWARN` or `PERROR`.

The debug level can be specified in the `DEBUG` environment variable.
To enable GDB debugging, set `GDB=1`.