netdev: Add IP address to `netdev`
Each network device is now assigned an IP address. This allows passing
the IP address that a frame was received on to the upper layers above
the drvier. Each network device could be assigned a differnet IP address
although, at present, it's just one default IP addres for all devices
(and in reality there is only one device).