netdev: Update to use input queue and `struct send_buf`
The netdev subsystem now features a ring buffer-based input queue
that stores packets received by a driver. When the devices fires
a receive interrupt, the packet is added to the queue and then
the interrupt handler is completed. This way, processing can
happen completely outside of the interrupt context.

The netdev subsystem is also switched to use `struct send_buf` for
transmitting data. This way, packets can be built-up with few
copies and only a single copy of the entire data happens when copying
the send buffer into an internal queue of a device.