netdev: Add warning when more than once device has the same IP
Having more than one device with the same IP address is not
really an issue we need to deal with yet, as we only have
a single device. The policy for assigning IP addresses and
picking devices can only conclusively be decided when there
are more devices (otherwise, it's likely that I would forget
to consider something).