sched: Add cooperate scheduling and `sleep` function
The `sched` module implements a simple cooperative scheduling
system that can run different tasks. Tasks can be added to the
scheduler with `sched_create_task`. Every time `sleep_ms` is
called, the scheduler checks which task should run next (based
on previous `sleep_ms` and `sched_create_task` calls) and
executes another taks.

The policy for scheduling is: none. The next task is simply
chosen from a doubly linked list that obeys no special ordering.
So, if you insist on using a technical term, this is closest to
a round robin scheduler.