ip: Guard header layout with `#if SYSTEM_BYTE_ORDER == NET_BYTE_ORDER`
If the system byte order is the same as the network byte order, then
the layout of the bit fields will be in the same order as in the RFCs.
Before, we compared if `SYSTEM_BYTE_ORDER == ORDER_LITTLE_ENDIAN` where
the order of the bit fields is flipped compared to the RFCs. Starting
with the layout for network byte order is easier to understand.