ip: Add functions to iterate the checksum calculation
These functions will be useful in the TCP implementation
where the internet checksum must be computed over the
pseudo header and the segment which are not contiguous
in memory. Instead of calling `internet_checksum` once,
one can call `internet_checksum_iterate` for each consecutive
region and then `internet_checksum_finalize` to apply the
final one's complement.