tcp: Purge TIME_WAIT connections after wait period has elapsed
A connection is put in the TIME_WAIT state after it was closed,
waiting for some time to elapse to retransmit FINs if they were
lost. We save the current time when moving a connection to the
TIME_WAIT state and re-use the connection only when the period
has elapsed. The purging happens in the `tcp_alloc_conn` and
`tcp_lookup_conn` functions because the connection can't be
re-used before one of these functions is called anyway.