ip, netdev: Add network device MTU
We need this so that TCP can fragment segments before
passing them on to the IP layer. Fragmenting at the
TCP layer is better than doing it at the IP layer because
it allows for fine-grained retransmission. If only a single
IP fragment of a large segment would be lost, the entire
fragment would need to be retransmitted.

On the other hand, we don't support retransmission yet so
this doesn't make a practical difference. But implementing
fragmentation at the IP layer would have required some
complicated code because the IP layer only receives a
`struct send_buf`.