tcp, web: Handle the peer closing the conn first
This means implementing the CLOSE_WAIT and LAST_ACK
states in the TCP module. It also means that we now
have a flag in the `tcp_conn_close` function that
indicates to the caller if the peer closed the connection.
Normally, the EOF character is used for this, which is
nice when TCP streams and files are meant to look the same,
but we can be more explicit.