web: Handle insufficient storage error
We don't expect to receive large requests since we just want
to server static pages (using GET) with short URLs. This means
we can use a small buffer for reception. But if a client sends
a request that's bigger than the buffer, they should get an error.

That's why now we're responding to requests bigger than the receive
buffer with a 507 Insufficient Storage error.