web: Retry until full header was received
Before, we only polled the TCP receive function to wait
for data to arrive. But this loop would return after
receiving just a single segment. This works if the initial
segment contains the entire header (which is often the case),
but the logic breaks down if someone transmits a header bigger
than the initial segment.

Now we have a second loop round the initial one that checks if
the '\r\n\r\n' sequence appears in the buffer and retries a few
more times to receive additional segments.

Since we only implement the GET method for static pages right
now, we don't need to try too often. The GET headers aren't
bigger than a few hundred bytes usually. But we still want to
handle the case where bigger headers arrive.