bootloader: Fix IDE polling and disable interrupts
We didn't wait correctly in the polling logic before.

The interrupts need to be disabled to avoid getting spurious
interrupts later. If these spurious interrupts aren't acknowledged,
it will block further interrupts (bad). But in the interrupt logic,
how do you tell appart spurious acknowledgments than should be
acknowledged from those that are erroneous and should lead to a
crash? That's a hard question. But by disabling interrupts in the
IDE driver, we can avoid dealing with it. (We also didn't need the
IDE interrupts for anything.)