init,config: Space for kernel image
We store the rootfs image in the kernel image which means that
kernel images get pretty big if we want to serve more data from
the web server. This commit creates more space for the kernel
image in the address range between KERN_BASE_PADDR and
KERN_DYN_VADDR (their difference is how much space is available
for the kernel image).

We need to allocate more page table pages in the .entry section
to facilitate this because the memory region mapped by the initial
page table must cover the entire kernel image plus the region of
dynamically managed memory that contains the page tables for the
final runtime page table. These page tables live in memory right
after the kernel image.