list: Fix TCP accept queue bug
New connections are added to the listen queue of the listen
connection that was used to create them. Once in the established
state, such connections can be accepted by the TCP API user. By
accepting a connection, it's removed from the accept queue.
Connections are also, unconditionally, removed from the accept
queue when they are deleted to make sure that no dangling connections
are ever in the accept queue.

Now here is a bug I experienced: some connections were added to the
accept queue of a listen connection. Then a connection in the TIME_WAIT
state would be deleted (this connection was removed from the accept
queue when it was accepted). Now, suddenly, the new connections from
the current accept queue had disappeared.

The problem: I had assumed that `dlist_remove` would remove the given
node from the list (it did) and that it would set the removed node
to be emtpy (it didn't!). The removed node would simply stay in the
same state that it was before. This explains how the bug could occur:
when the TIME_WAIT connection was originally accepted, it had beed
removed from the accept queue, but the accept queue field in this
connection didn't change. So when the TIME_WAIT connection was finally
deleted, calling `dlist_remove` on it again reset the queue to the
state from before when the TIME_WAIT connection had originally been
removed.

We solve this by making `dlist_remove` reset the given node to be
an empty list. This way, calling `dlist_remove` on the same node
twice works without issues.