tcp: Keep right edge of receive window fixed (avoid SWS)
This commit is based on my understanding of section
3.8.6.2.2 of RFC 9293. It's about avoiding Silly
Window Syndrome by keeping the right edge of the receive
window fixed over longer durations. We do this by
updating window field in the TCP header so that
RCV.NXT + RCV.WND are constant over some duration
of time.

I didn't really observe this behavior in the TCP
traffic generated by Linux or my iPhone so I'm not
sure if my interpretation of the text is correct.
I may revert this commit later.