tcp: Rename `struct circ_buf` to `struct recv_buf`
I originally intended `circ_buf` to be used both on the
receive and on the transmit side. It turned out that using
a circular buffer on the transmit side doesn't make sense:
it's much easier to keep a queue of send buffer structures,
one for each segment that can be retransmitted. This way,
we can easily associate the right metadata with each segment
that can be retransmitted. If we had put all data that we
might need to retransmit in a single contigous circ_buf,
tracking all the extra state around each of the segments in
it would have been complicated and error-prone.

So I chose to rename `circ_buf` to `recv_buf` because that's
what it actually is in the code: a structure only used to buffer
received data.