tcp: Await ACKs for retransmissions
This commit fixes multiple issues with retransmissions. First,
retransmissions often happended in the `FIN_WAIT_2` state. A
TCP connection ends up in this state if the connection is
half open because our server closed it. If the server transmitted
a large amount of data, there is likely still some data waiting
for ACKs from the peer. If no ACKs arrive, it's retransmitted or
eventually discarded. But the `FIN_WAIT_2` handler didn't update
the state of the connection if new ACKs arrived. Thus, no ACKs
could arrive.

Second, we now (correctly) compare ACK numbers with "greater than
or equal to" in the check if we can delete a fragment in the SBQ.
This used "greater than" before but that's wrong.

Third, we leave the retransmission loop of a single connection
as soon as the current node in the SBQ has a required ACK number
greater than the biggest ACK number that has arrived up to this
point. Before leaving, though, we check if the current node's
retransmission timer has expired. If it has, we can retransmit
it. This change as two effects: (1) new ACKs from the peer can
arrive after leaving the retransmission loop so that we can
actually advance in the retranmission queue by removing old
fragments that have arrived at the peer; (2) the retransmission
loop stops on the first fragment that hasn't been acknowledged
yet. The latter change is very important to reduce excess traffic.
Imagine you have fragments A, B and C in the SBQ. If fragment A
hasn't yet been acknowledged, it doesn't make sense to retransmit
fragments B and C, even if the retransmission timers of the latter
two have expired.

With the three fixes above, retransmission works pretty well now.
A few fragments are retransmitted and in between new ACKs arrive
which leads the TCP subsystem to remove already retransmitted
fragments from the SBQ and allows the retranmission to progress.