tcp: Fix connection leaks
This commit fixes the mistake of only handing out
connections to users when they are in the ESTABLISHED
state. With the current system of connection states,
it becomes the user's responsiblitiy to close a connection
once the connection enters the ESTABLISHED state. From
there, the connection can get to the CLOSE_WAIT and the
RESET states without the user taking any action (it
enters other states when the user calls close on it).
With timing is right, a connection can be moved from
the ESTABLISHED state to the CLOSE_WAIT or RESET state
before the user had a chance to access it. In such a
case, the connection wouldn't be freed because the
user can't access it but the TCP subsystem assumes
that the connection is freed by the user. Now, all
connections in the ESTABLISHED, CLOSE_WAIT and RESET
states can be accessed by the user.