tcp: Improve `tcp_lookup_conn` function
This commit removes the `use_wilcard` flag from the `tcp_lookup_conn`
function. The function prefers exact matches, but, if there are none,
it will fall back to returning a wildcard match, if any.

The new implementation works for both of the original call sites. It
relies on the fact that there is at most one exact match for all
connections.