isr, idt, init, time: Clean up IRQ management
All IRQs are disabled by default now. To set up interrupts
and IRQs, `interrupt_init` is called first which disables
are IRQs and remaps the PIC to set up the cascade. Now,
when a interrupt handler for an IRQ is registered with
`isr_register_handler`, the IRQ is explicity enabled.

The upside here is that concerns are separated much better.
E.g., the PIT timer interrupt handler no longer needs to be
registers before setting up and enabling interrupts, simply
because the PIT IRQ isn't enabled by default any longer.